@@ -11,15 +11,15 @@
   },
   "dependencies": {
     "heft-web-rig-library-tutorial": "workspace:*",
-    "react": "~17.0.2",
-    "react-dom": "~17.0.2",
+    "react": "~19.2.3",
+    "react-dom": "~19.2.3",
     "tslib": "~2.8.1"
   },
   "devDependencies": {
     "@rushstack/heft": "workspace:*",
     "@rushstack/heft-web-rig": "workspace:*",
-    "@types/react": "17.0.74",
-    "@types/react-dom": "17.0.25",
+    "@types/react": "19.2.7",
+    "@types/react-dom": "19.2.3",
     "@types/webpack-env": "1.18.8",
     "eslint": "~9.37.0",
     "local-eslint-config": "workspace:*"
