@@ -13,15 +13,15 @@
     "_phase:test": "heft run --only test -- --clean"
   },
   "dependencies": {
-    "react": "~17.0.2",
-    "react-dom": "~17.0.2",
+    "react": "~19.2.3",
+    "react-dom": "~19.2.3",
     "tslib": "~2.8.1"
   },
   "devDependencies": {
     "@rushstack/heft-web-rig": "workspace:*",
     "@rushstack/heft": "workspace:*",
-    "@types/react-dom": "17.0.25",
-    "@types/react": "17.0.74",
+    "@types/react-dom": "19.2.3",
+    "@types/react": "19.2.7",
     "@types/webpack-env": "1.18.8",
     "eslint": "~9.37.0",
     "local-eslint-config": "workspace:*"
