@@ -10,8 +10,8 @@
     "_phase:test": "heft run --only test -- --clean"
   },
   "dependencies": {
-    "react-dom": "~17.0.2",
-    "react": "~17.0.2",
+    "react-dom": "~19.2.3",
+    "react": "~19.2.3",
     "tslib": "~2.8.1"
   },
   "devDependencies": {
@@ -21,8 +21,8 @@
     "@rushstack/heft-webpack5-plugin": "workspace:*",
     "@rushstack/heft": "workspace:*",
     "@types/heft-jest": "1.0.1",
-    "@types/react": "17.0.74",
-    "@types/react-dom": "17.0.25",
+    "@types/react": "19.2.7",
+    "@types/react-dom": "19.2.3",
     "@types/webpack-env": "1.18.8",
     "css-loader": "~6.6.0",
     "eslint": "~9.37.0",
