@@ -21,8 +21,8 @@
     "@rushstack/webpack4-module-minifier-plugin": "workspace:*",
     "@types/heft-jest": "1.0.1",
     "@types/node": "20.17.19",
-    "@types/react-dom": "17.0.25",
-    "@types/react": "17.0.74",
+    "@types/react-dom": "19.2.3",
+    "@types/react": "19.2.7",
     "@types/webpack-env": "1.18.8",
     "autoprefixer": "~10.4.2",
     "css-loader": "~5.2.7",
@@ -30,8 +30,8 @@
     "html-webpack-plugin": "~4.5.2",
     "postcss-loader": "~4.1.0",
     "postcss": "~8.4.6",
-    "react-dom": "~17.0.2",
-    "react": "~17.0.2",
+    "react-dom": "~19.2.3",
+    "react": "~19.2.3",
     "style-loader": "~2.0.0",
     "typescript": "~5.8.2",
     "webpack": "~4.47.0"
