@@ -18,22 +18,22 @@
     "start": "heft start"
   },
   "dependencies": {
-    "@fluentui/react-components": "~9.27.0",
-    "@fluentui/react": "^8.96.1",
-    "@reduxjs/toolkit": "~1.8.6",
-    "react-dom": "~17.0.2",
-    "react-hook-form": "~7.24.1",
-    "react-redux": "~8.0.4",
-    "react": "~17.0.2",
-    "redux": "~4.2.0",
+    "@fluentui/react-components": "~9.72.9",
+    "@fluentui/react": "~8.125.3",
+    "@reduxjs/toolkit": "~2.11.2",
+    "react-dom": "~19.2.3",
+    "react-hook-form": "~7.69.0",
+    "react-redux": "~9.2.0",
+    "react": "~19.2.3",
+    "redux": "~5.0.1",
     "scheduler": "0.19.0",
     "tslib": "~2.8.1"
   },
   "devDependencies": {
     "@rushstack/heft": "workspace:*",
     "@rushstack/ts-command-line": "workspace:*",
-    "@types/react": "17.0.74",
-    "@types/react-dom": "17.0.25",
+    "@types/react": "19.2.7",
+    "@types/react-dom": "19.2.3",
     "@types/react-redux": "~7.1.22",
     "@types/vscode": "1.103.0",
     "eslint": "~9.37.0",
