@@ -2,7 +2,7 @@
 // See LICENSE in the project root for license information.
 
 import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import * as ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 
 import { App } from './App';
@@ -10,9 +10,8 @@ import './start.css';
 import { store } from './store';
 
 const rootDiv: HTMLElement = document.getElementById('root') as HTMLElement;
-ReactDOM.render(
+ReactDOM.createRoot(rootDiv).render(
   <Provider store={store}>
     <App />
-  </Provider>,
-  rootDiv
+  </Provider>
 );
