@@ -2,12 +2,13 @@
 // See LICENSE in the project root for license information.
 
 import * as React from 'react';
+import type { FieldErrors } from 'react-hook-form';
 
 export interface IErrorMessageProps {
-  message?: string;
+  message?: FieldErrors[string] | string;
 }
 
-export const ErrorMessage = ({ message }: IErrorMessageProps): JSX.Element => {
+export const ErrorMessage = ({ message }: IErrorMessageProps): React.ReactElement => {
   // eslint-disable-next-line no-console
   console.log('ErrorMessage...', message);
   return message ? (
